Add new Question and Comment working
Users can now add questions and reply to them. Click the button at the
top to ask a new question. Click on a "X Comments" to view the comments
for the question. Still not quite done yet with everything I wanted to
do (like the Back button and kicking the user off the comments when the
question expires).